Develop header search component
